requirements.txt
aiofiles==0.7.0
alembic==1.7.4
arrow==1.2.0
asyncpg==0.24.0
bcrypt==3.2.0
broadcaster==0.2.0
celery==5.1.2
emails==0.6
fastapi==0.70.0
httptools==0.3.0
jupyter==1.0.0
passlib==1.7.4
psycopg2-binary==2.9.1
pydantic[email]==1.8.2
python-jose==3.3.0
python-multipart==0.0.5
SQLAlchemy==1.4.25
tenacity==8.0.1
uvicorn==0.15.0
websockets==10.0